requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
distro==1.9.0
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pydantic==2.10.4
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
shiboken6==6.8.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
